requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dj-database-url==3.0.1
Django==5.0.6
django-browser-reload==1.12.1
django-cors-headers==4.4.0
django-tailwind==3.8.0
djangorestframework==3.15.2
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
frozendict==2.4.4
gast==0.5.5
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
idna==3.4
ImageHash==4.3.1
imutils==0.5.4
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.1.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.4.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
ml-dtypes==0.3.2
multimethod==1.11.2
multitasking==0.0.11
namex==0.0.8
narwhals==1.44.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
notebook==7.1.3
notebook_shim==0.2.3
numpy==1.26.3
openai==1.31.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pandas-datareader==0.10.0
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
peewee==3.17.5
phik==0.12.4
pillow==10.2.0
platformdirs==3.11.0
plotly==6.2.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.6.0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.12.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.1
Send2Trash==1.8.2
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.35.0
style==1.1.0
tangled-up-in-unicode==0.2.0
tdameritrade==0.2.1
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.11.0
tzdata==2023.4
update==0.0.1
uri-template==1.3.0
urllib3==2.0.7
uv==0.1.44
virtualenv==20.26.3
visions==0.7.4
watchdog==4.0.1
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.9.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xgboost==3.0.2
yfinance==0.2.40
